pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_context_studio"
version = "1.1.0"
description = "Desktop Application per la generazione di Knowledge Base per Agenti AI"
authors = [
  { name="Giancarlo Allegretti" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "customtkinter>=5.2.0",
    "google-generativeai>=0.5.0",
    "cryptography>=41.0.0",
]

[project.scripts]
ai-context-studio = "ai_context_studio.main:main"

[tool.setuptools.packages.find]
where = ["src"]
